Link to assignment in README and disclaim six-week time limit
So that if someone looks at the game and thinks it looks unpolished, they won’t think it’s because we didn’t notice the problems with it.

Six weeks is from when we were given the assignment to think of a game to when the final version was due.