Make preloading progress update constantly
It now updates even when a file has not finished loading. This might add more feedback for those on slow connections; the number will gradually go up instead of holding and then jumping up a lot.

I update at 4 FPS, a slow amount, so that the user has time to read the numbers if they go up somewhat fast.